+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This is a list of changes to `stac-geoparquet`.
 ## Unreleased
 
 - Don't configure logging level at module level (<https://github.com/stac-utils/stac-geoparquet/pull/152>)
+- Write null datetime columns in UTC to support TZ-aware comparisons with other columns (<https://github.com/stac-utils/stac-geoparquet/pull/157>)
 
 ## 0.8.1